This commit fixes #2586.

+++ src/Control/ASV/HeadingAndSpeed/Task.cpp
@@ -83,14 +83,14 @@ namespace Control
         Delta m_last_estate;
         // Motor Thrust.
         IMC::SetThrusterActuation m_motor[2];
-        //! Control loops last reference time
-        float m_scope_ref;
+        //! Control loops last reference
+        uint32_t m_scope_ref;
         // Task arguments.
         Arguments m_args;
 
         Task(const std::string& name, Tasks::Context& ctx):
           Tasks::Task(name, ctx),
-          m_scope_ref(0.0)
+          m_scope_ref(0)
         {
           param("Maximum Motor Command", m_args.max_motor)
           .defaultValue("1.0")
